[PL-20070]: single instance of Managed Resource Group (#27508)
* [PL-20070]: initial changes

* [PL-20070]: keep old

* [PL-20070]: compile

* [PL-20070]: new variable

* [PL-20070]: new variable

* [PL-20070]: checkstyle

* [PL-20070]: clang

* [PL-20070]: test

* [PL-20070]: variable

* [PL-20070]: allowed scope levels

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: mapper

* [PL-20070]: aeriform

* [PL-20070]: aeriform

* [PL-20070]: remove

* [PL-20070]: migration

* [PL-20070]: testing

* [PL-20070]: testing

* [PL-20070]: job test

* [PL-20070]: service test

* [PL-20070]: service test

* [PL-20070]: filter

* [PL-20070]: filter

* [PL-20070]: json ignore

* [PL-20070]: comments

* [PL-20070]: comments

* [PL-20070]: get test

* [PL-20070]: delete managed test

* [PL-20070]: update test

* [PL-20070]: update test

* [PL-20070]: sync job tests

* [PL-20070]: checkstyle

* [PL-20070]: test fix

* [PL-20070]: bazel

* [PL-20070]: redis

* [PL-20070]: redis

+++ 925-access-control-service/src/main/java/io/harness/accesscontrol/resources/resourcegroups/events/ResourceGroupEventHandler.java
@@ -1,5 +1,6 @@
 package io.harness.accesscontrol.resources.resourcegroups.events;
 
+import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.eventsframework.EventsFrameworkMetadataConstants.ACTION;
 import static io.harness.eventsframework.EventsFrameworkMetadataConstants.DELETE_ACTION;
 
@@ -51,6 +52,9 @@ public class ResourceGroupEventHandler implements EventHandler {
               .orgIdentifier(stripToNull(resourceGroupEntityChangeDTO.getOrgIdentifier()))
               .projectIdentifier(stripToNull(resourceGroupEntityChangeDTO.getProjectIdentifier()))
               .build();
+      if (isEmpty(scopeParams.getAccountIdentifier())) {
+        return true;
+      }
       Scope scope = ScopeMapper.fromParams(scopeParams);
       if (getEventType(message).equals(DELETE_ACTION)) {
         harnessResourceGroupService.deleteIfPresent(stripToNull(resourceGroupEntityChangeDTO.getIdentifier()), scope);
